requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
bcrypt
firebase-admin
google-cloud-firestore
python-jose[cryptography]
gunicorn
jinja2
